+++ board/recalbox/fsoverlay/recalbox/scripts/recalbox-config.sh
@@ -635,6 +635,12 @@ if [[ "$command" == "updateesinput" ]]; then
     exit $?
 fi
 
+if [[ "$command" == "getEmulatorDefaults" ]]; then
+    emulator="$mode"
+    python -c "import json; import sys; sys.path.append('/usr/lib/python2.7/site-packages/configgen'); from emulatorlauncher import emulators; print(json.dumps(emulators['$emulator'].config))"
+    exit 0
+fi
+
 echo "Uknown command $command"
 recallog -e "recalbox-config.sh: unknown command $command"